+++ CHANGELOG.rst
@@ -1,3 +1,12 @@
+Next release
+------------
+
+:Date: July X, 2024
+
+* `@marek-serafin <https://github.com/stoprocent>`__ : Fixed typo in setting peer_cid while processing message 1.
+* `@marek-serafin <https://github.com/stoprocent>`__ : Fixed zcbor. Added a method to avoid generation of duplicated types.
+* `@kamil-kielbasa <https://github.com/kamil-kielbasa>`__ : Added EDHOC error message compose & process with unit tests.
+
 Version 0.3.0
 -------------